miniengine: prepare for refactoring and cleanups
