+++ internal/miniengine/callbacks.go
@@ -6,8 +6,10 @@ package miniengine
 
 import "github.com/ooni/probe-cli/v3/internal/model"
 
-// callbacks implements [model.ExperimentCallbacks]
+// callbacks implements [model.ExperimentCallbacks] and emits
+// the callbacks events using the given channel.
 type callbacks struct {
+	// emitter is the channel where to emit events.
 	emitter chan<- *Event
 }
 
@@ -20,6 +22,7 @@ func (c *callbacks) OnProgress(progress float64, message string) {
 		Message:   message,
 		Progress:  progress,
 	}
+	// Implementation note: it's fine to lose interim events
 	select {
 	case c.emitter <- event:
 	default:
